+++ src/main/java/com/hiutaleapp/service/LocationService.java
@@ -2,9 +2,18 @@ package com.hiutaleapp.service;
 
 import com.hiutaleapp.dto.LocationDTO;
 import com.hiutaleapp.entity.Location;
+import com.hiutaleapp.entity.User;
 import com.hiutaleapp.repository.LocationRepository;
+import com.hiutaleapp.util.errors.DataViolationException;
+import com.hiutaleapp.util.errors.DatabaseConnectionException;
+import com.hiutaleapp.util.forms.LocationForm;
+import com.hiutaleapp.util.errors.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessResourceFailureException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.CannotCreateTransactionException;
 
 import java.util.List;
 import java.util.Optional;
@@ -25,8 +34,23 @@ public class LocationService {
         return locationRepository.findById(id).map(this::mapToDTO);
     }
 
-    public LocationDTO createLocation(Location location) {
-        return mapToDTO(locationRepository.save(location));
+    public LocationDTO createLocation(Long userId, LocationForm locationForm) {
+        try {
+            Location location = new Location();
+            User user = new User();
+
+            user.setUserId(userId);
+            location.setName(locationForm.getName());
+            location.setAddress(locationForm.getAddress());
+            location.setCity(locationForm.getCity());
+            location.setPostalCode(locationForm.getPostal());
+            location.setCreator(user);
+            return mapToDTO(locationRepository.save(location));
+        } catch (CannotCreateTransactionException e) {
+            throw new DatabaseConnectionException("Could not connect to the database");
+        } catch (DataIntegrityViolationException e) {
+            throw new DataViolationException("Could not create location, similar location already exists"); // Placeholder
+        }
     }
 
     public LocationDTO updateLocation(Long id, Location location) {
@@ -34,8 +58,24 @@ public class LocationService {
         return mapToDTO(locationRepository.save(location));
     }
 
-    public void deleteLocation(Long id) {
-        locationRepository.deleteById(id);
+    public Boolean deleteLocation(Long userId, Long id) {
+        try {
+            Optional<LocationDTO> location = getLocationById(id);
+            if (location.isPresent()) {
+                if (location.get().getCreatorId() == userId) {
+                    locationRepository.deleteById(id);
+                    return true;
+                } else {
+                    throw new AuthorizationDeniedException("You do not have permission to delete this location");
+                }
+            } else {
+                throw new NotFoundException("Location with this ID does not exist");
+            }
+        } catch (DataAccessResourceFailureException e) {
+            throw new DatabaseConnectionException("Could not connect to the database");
+        } catch (DataIntegrityViolationException e) {
+            throw new DataViolationException("Could not delete this location because some events are dependent upon it");
+        }
     }
 
     public LocationDTO mapToDTO(Location location) {
